oc checker code refactor
